fix(back) changing status db

+++ src/main/java/com/grupo3/consumer/dao/OrderDAO.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.grupo3.consumer.model.Order;
 
-public interface OrderDAO extends JpaRepository<Order, Integer> {
-    public Optional<Order> findById(Integer id);
+public interface OrderDAO extends JpaRepository<Order, Long> {
+    public Optional<Order> findById(Long id);
 }
