+++ src/main/java/com/grupo3/consumer/services/IOrderService.java
@@ -3,9 +3,9 @@ package com.grupo3.consumer.services;
 import com.grupo3.consumer.model.Order;
 
 public interface IOrderService {
-    public boolean orderExists(Integer orderId);
+    public boolean orderExists(Long orderId);
 
-    public boolean orderInProgress(Integer orderId);
+    public boolean orderInProgress(Long orderId);
 
-    public Order updateStatusOrderSuccess(Integer orderId);
+    public Order updateStatusOrderSuccess(Long orderId);
 }
